+++ src/components/OfflineMessage.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react'
 import { Pressable, Text, View } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
-import { colors } from '~/styles/colors'
+import { colors } from '~/features/ui/theme/colors'
 
 export const OfflineMessage = () => {
   const { bottom } = useSafeAreaInsets()
