+++ src/features/versionCheck/components/ForcedUpdate.tsx
@@ -1,10 +1,10 @@
 import * as Linking from 'expo-linking'
 import { useEffect } from 'react'
-import { View, Alert } from 'react-native'
+import { View, Alert, AlertButton } from 'react-native'
 
 import { getStoreLink } from '~/utils/getStoreLink'
 
-export const ForcedUpdate = () => {
+export const ForcedUpdate = ({ cancelable = true }: { cancelable?: boolean }) => {
   useEffect(() => {
     const handleUpdate = async () => {
       const storeLink = getStoreLink()
@@ -16,18 +16,23 @@ export const ForcedUpdate = () => {
       }
     }
 
+    const buttonCancel = {
+      text: 'Not now',
+      onPress: () => {},
+    }
+
+    const buttonUpdate = {
+      text: 'Update',
+      onPress: handleUpdate,
+    }
+
     Alert.alert(
       'Update Available',
       'Updating to the latest version provides the newest features, security updates, and bug fixes. Tap below to update the app.',
-      [
-        {
-          text: 'Update',
-          onPress: handleUpdate,
-        },
-      ],
-      { cancelable: false },
+      [cancelable && buttonCancel, buttonUpdate].filter(Boolean) as AlertButton[],
+      { cancelable },
     )
-  }, [])
+  }, [cancelable])
 
   return <View />
 }
